Treat them as separate projects instead of modules in one project
This seems a better approach as it seems to make the get-up-and-running process from the freshly cloned repo more robust and the project default location is in the projects directly and not in the OntoEngine-root. I noticed this when setting up Jena TDB and it created the persistence directory in root by default instead of the actual project folder where I want it to be by default.
Plus let's ignore all of the .idea/ folder for now.