Use Jena TDB for persistence, get directory from properties file
In the JenaController constructor, tdbDir didn't get the value from the properties file assigned yet. But when moving the initialization into a @PostConstruct method, Spring had time to collect that value. @PreDestroy ensures that the model will be closed when shutting down the server. Not sure if that is the case in all crashing scenarios though.