Create new edges by dragging nodes close to each other
The drag-gesture is occupied by the force-graph library, couldn't easily override/augment that with own behaviour. Also tried mouse-wheel click-drag but didn't get it to work. So I thought it'd be fun to drag nodes close to each other to signal the creation of a new edge pointing from the dragged one to the bumped-into one :)